requirements.txt
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0
PyQt5-sip>=12.12.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy<2
requests>=2.31.0
ollama
